fix: add permission requirements to new endpoints
