+++ backend/licensing/rest/properties.py
@@ -14,7 +14,7 @@ from licensing.models import (
     LicenseStatusChoices,
     Species,
 )
-from .utils import LabeledChoiceViewset
+from .utils import LabeledChoiceViewset, DjangoProtectedModelPermissions
 
 
 class ActorSerializer(serializers.ModelSerializer):
@@ -80,21 +80,29 @@ class PermissionPropertySerializer(serializers.ModelSerializer):
 
 
 class ActorViewSet(viewsets.ModelViewSet):
+    permission_classes = [DjangoProtectedModelPermissions]
+
     queryset = Actor.objects.all()
     serializer_class = ActorSerializer
 
 
 class SpeciesViewSet(viewsets.ModelViewSet):
+    permission_classes = [DjangoProtectedModelPermissions]
+
     queryset = Species.objects.all()
     serializer_class = SpeciesSerializer
 
 
 class PermissionTypeViewSet(viewsets.ModelViewSet):
+    permission_classes = [DjangoProtectedModelPermissions]
+
     queryset = LicensePermissionType.objects.all()
     serializer_class = PermissionTypeSerializer
 
 
 class PermissionPropertyViewSet(viewsets.ModelViewSet):
+    permission_classes = [DjangoProtectedModelPermissions]
+
     queryset = LicensePermissionProperty.objects.all()
     serializer_class = PermissionPropertySerializer
